Conflicts:
	articleflow/views.py

+++ articleflow/views.py
@@ -261,7 +261,7 @@ class ArticleDetailMain(View):
         return render_to_response(self.template_name, context, context_instance=RequestContext(request))
 
     def get_context_data(self, kwargs):
-        article = Article.objects.get(doi=kwargs['doi'])
+        article = get_object_or_404(Article, doi=kwargs['doi'])
         context = ({
                 'article': article,
             })
@@ -283,11 +283,18 @@ class ArticleDetailTransitionPanel(View):
 
 class ArticleDetailTransitionUpload(View):
 
-    def handle_uploaded_file(self, f, destination_pathname):
-        logger.debug("Writing uploaded file to %s" % destination_pathname)
-        with open(destination_pathname, 'wb+') as destination:
-            for chunk in f.chunks():
-                destination.write(chunk)
+    def handle_uploaded_file(self, f, destinations, filename):
+        logger.debug("Handling file upload for %s" % filename)
+        dests = destinations.split(' ')
+        for dest in dests:
+            if not os.path.exists(dest):
+                logger.error("File upload destination path, %s, does not exist. skipping." % dest)
+                continue
+            destination_pathname = os.path.join(dest, filename)
+            logger.debug("Writing uploaded file to %s" % destination_pathname)
+            with open(destination_pathname, 'wb+') as destination:
+                for chunk in f.chunks():
+                    destination.write(chunk)
 
     def post(self, request, *args, **kwargs):
         if not request.user.is_authenticated():
@@ -302,7 +309,7 @@ class ArticleDetailTransitionUpload(View):
             if transition in article.possible_transitions():
                 logger.debug("Handling uploaded file: %s . . ." % request.FILES['file'].name)
                 f_name, extension = os.path.splitext(request.FILES['file'].name)
-                self.handle_uploaded_file(request.FILES['file'], os.path.join(transition.file_upload_destination, "%s%s" % (article.doi, extension)))
+                self.handle_uploaded_file(request.FILES['file'], transition.file_upload_destination, "%s%s" % (article.doi, extension))
                 article.execute_transition(transition, request.user)
                 return HttpResponseRedirect(reverse('detail_main', args=(article.doi,)))
 
@@ -665,6 +672,8 @@ def send_file(pathname, attachment_name=None):
         raise IOError()
     logger.debug("Opening file at '%s' for reading." % file_name[0])
     wrapper = FileWrapper(file(file_name[0]))
+    mime, enc = mimetypes.guess_type(file_name[0], False)
+    logger.debug("%s mimetype: %s, encoding: %s" % (file_name[0], mime, enc))
     throwaway, extension = os.path.splitext(file_name[0])
     response = HttpResponse(wrapper, content_type=mime)
     response['Content-Length'] = os.path.getsize(file_name[0])
